requirements.txt
fastapi
uvicorn[standard]
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langsmith==0.1.48
openai==1.20.0
python-dotenv==1.0.1
langchain-text-splitters
pypdf
transformers
bs4
langchain_astradb
bson